Zmiana nazwy modułu z inventory na backend
