+++ backend/pom.xml
@@ -9,9 +9,9 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>org.polsl</groupId>
-	<artifactId>inventory</artifactId>
+	<artifactId>backend</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>inventory</name>
+	<name>backend</name>
 	<description>Demo project for Spring Boot</description>
 
 	<properties>
