+++ backend/src/main/java/org/polsl/backend/entity/Hardware.java
@@ -1,4 +1,4 @@
-package org.polsl.inventory.entity;
+package org.polsl.backend.entity;
 
 import javax.persistence.*;
 import java.util.Set;
