+++ backend/src/main/java/org/polsl/backend/entity/Software.java
@@ -1,4 +1,4 @@
-package org.polsl.inventory.entity;
+package org.polsl.backend.entity;
 
 import javax.persistence.*;
