+++ backend/src/main/java/org/polsl/backend/repository/AffiliationRepository.java
@@ -1,6 +1,6 @@
-package org.polsl.inventory.repository;
+package org.polsl.backend.repository;
 
-import org.polsl.inventory.entity.Affiliation;
+import org.polsl.backend.entity.Affiliation;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
