+++ backend/src/main/java/org/polsl/backend/repository/HardwareRepository.java
@@ -1,6 +1,6 @@
-package org.polsl.inventory.repository;
+package org.polsl.backend.repository;
 
-import org.polsl.inventory.entity.Hardware;
+import org.polsl.backend.entity.Hardware;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
